+++ tests/unit/test_setup.py
@@ -1,7 +1,8 @@
 from os.path import abspath, dirname, isfile, join
 from shutil import rmtree
 from subprocess import DEVNULL, call
-from unittest import TestCase
+
+import pytest
 
 from issuewatcher import APPLICATION_NAME, __version__ as app_version
 
@@ -14,30 +15,37 @@ BUILD_ARTEFACTS = [
     for folder in ["dist", "build", UNDERSCORED_APPLICATION_NAME + ".egg-info"]
 ]
 
+# pylint: disable=redefined-outer-name
+
 
 def cleanup_build_artefacts():
     for folder in BUILD_ARTEFACTS:
         rmtree(folder, ignore_errors=True)
 
 
-class BuildProcess(TestCase):
-    def setUp(self):
-        cleanup_build_artefacts()
-        self._build_return_code = call(
+@pytest.fixture(scope="session")
+def build_return_code():
+    cleanup_build_artefacts()
+
+    try:
+        yield call(
             ["python", "setup.py", "sdist", "bdist_wheel"], stdout=DEVNULL, stderr=DEVNULL
         )
-        self.addCleanup(cleanup_build_artefacts)
+    finally:
+        cleanup_build_artefacts()
 
-    def test_it_builds_sources(self):
-        self.assertEqual(0, self._build_return_code)
 
-    def assert_file_is_build(self, suffix, replace_hyphens=False):
+class TestBuildProcess:
+    @staticmethod
+    def test_it_builds_sources(build_return_code):
+        assert build_return_code == 0
+
+    @staticmethod
+    def assert_file_is_build(suffix, replace_hyphens=False):
         app_name = UNDERSCORED_APPLICATION_NAME if replace_hyphens else APPLICATION_NAME
         filename = join(BUILD_ARTEFACTS[0], f"{app_name}-{app_version}{suffix}")
 
-        self.assertTrue(
-            isfile(filename), msg=f"File '{filename}' does not exist or is not a file."
-        )
+        assert isfile(filename), f"File '{filename}' does not exist or is not a file."
 
     def test_it_creates_whl_file(self):
         self.assert_file_is_build("-py3-none-any.whl", replace_hyphens=True)