+++ tests/unit/test_temporary_cache.py
@@ -1,14 +1,12 @@
 import os
 from time import time
 from typing import Dict, List, Union
-from unittest import TestCase
 from unittest.mock import patch
 
-from parameterized import parameterized
+import pytest
 from ujson import dumps, loads
 
 from temporary_cache import TemporaryCache
-from tests.helpers.parameterized import get_test_case_name_without_index
 
 _PROJECT = "radeklat/issue-watcher"
 _KEY_IN = "1"
@@ -37,46 +35,54 @@ def _get_instance() -> TemporaryCache:
     return TemporaryCache(_PROJECT)
 
 
-class TempCache(TestCase):
-    def test_it_can_read_from_a_file(self):
+class TestTempCache:
+    @staticmethod
+    def test_it_can_read_from_a_file():
         _create_temp_file({_PROJECT: {_KEY_OUT: [_VALUE, int(time())]}})
-        self.assertEqual(_get_instance()[_KEY_IN], _VALUE)
+        assert _get_instance()[_KEY_IN] == _VALUE
 
-    def test_it_can_write_into_a_file(self):
+    @staticmethod
+    def test_it_can_write_into_a_file():
         _create_temp_file()
         with patch("time.time", return_value=10):
             _get_instance()[_KEY_IN] = _VALUE
-        self.assertDictEqual(loads(_read_temp_file()), {_PROJECT: {_KEY_OUT: [_VALUE, 10]}})
+        assert loads(_read_temp_file()) == {_PROJECT: {_KEY_OUT: [_VALUE, 10]}}
 
-    def test_it_raises_key_error_when_key_is_missing(self):
+    @staticmethod
+    def test_it_raises_key_error_when_key_is_missing():
         _create_temp_file()
-        with self.assertRaises(KeyError):
+        with pytest.raises(KeyError):
             _value = _get_instance()[_KEY_IN]
 
-    def test_it_ignores_non_json_file(self):
+    @staticmethod
+    def test_it_ignores_non_json_file():
         _create_temp_file('asdjfkk{:}:""')
         with patch("time.time", return_value=10):
             _get_instance()[_KEY_IN] = _VALUE
-        self.assertDictEqual(loads(_read_temp_file()), {_PROJECT: {_KEY_OUT: [_VALUE, 10]}})
+        assert loads(_read_temp_file()) == {_PROJECT: {_KEY_OUT: [_VALUE, 10]}}
 
-    def test_it_ignores_json_file_without_top_level_dict(self):
+    @staticmethod
+    def test_it_ignores_json_file_without_top_level_dict():
         _create_temp_file([1, 2, 3, 4])
         with patch("time.time", return_value=10):
             _get_instance()[_KEY_IN] = _VALUE
-        self.assertDictEqual(loads(_read_temp_file()), {_PROJECT: {_KEY_OUT: [_VALUE, 10]}})
+        assert loads(_read_temp_file()) == {_PROJECT: {_KEY_OUT: [_VALUE, 10]}}
 
-    def test_it_ignores_missing_file(self):
+    @staticmethod
+    def test_it_ignores_missing_file():
         _remove_temp_file()
 
-        with self.assertRaises(KeyError):
+        with pytest.raises(KeyError):
             _value = _get_instance()[_KEY_IN]
 
-    def test_it_can_be_cleared(self):
+    @staticmethod
+    def test_it_can_be_cleared():
         _create_temp_file({_PROJECT: {_KEY_OUT: _VALUE}})
         _get_instance().clear()
-        self.assertDictEqual(loads(_read_temp_file()), {})
+        assert loads(_read_temp_file()) == {}
 
-    def test_it_separates_projects(self):
+    @staticmethod
+    def test_it_separates_projects():
         _remove_temp_file()
 
         class TestCache(TemporaryCache):
@@ -86,74 +92,88 @@ class TempCache(TestCase):
             TestCache(_PROJECT)[_KEY_IN] = _VALUE
             TestCache(_PROJECT + "x")[_KEY_IN] = _VALUE
 
-        self.assertDictEqual(
-            loads(_read_temp_file()),
-            {_PROJECT: {_KEY_OUT: [_VALUE, 10]}, _PROJECT + "x": {_KEY_OUT: [_VALUE, 10]}},
-        )
+        assert loads(_read_temp_file()) == {
+            _PROJECT: {_KEY_OUT: [_VALUE, 10]},
+            _PROJECT + "x": {_KEY_OUT: [_VALUE, 10]},
+        }
 
 
-class TempCacheGet(TestCase):
-    def setUp(self):
+class TestTempCacheGet:
+    @staticmethod
+    @pytest.fixture(autouse=True)
+    def set_up():
         _create_temp_file({_PROJECT: {_KEY_OUT: [_VALUE, int(time())]}})
 
-    def test_it_returns_value_if_exists(self):
-        self.assertEqual(_get_instance().get(_KEY_IN), _VALUE)
+    @staticmethod
+    def test_it_returns_value_if_exists():
+        assert _get_instance().get(_KEY_IN) == _VALUE
 
-    def test_it_returns_none_by_default_when_key_is_missing(self):
-        self.assertIsNone(_get_instance().get("2"))
+    @staticmethod
+    def test_it_returns_none_by_default_when_key_is_missing():
+        assert _get_instance().get("2") is None
 
-    def test_it_returns_custom_default_value_when_key_is_missing(self):
-        self.assertEqual(_get_instance().get("3", _VALUE), _VALUE)
+    @staticmethod
+    def test_it_returns_custom_default_value_when_key_is_missing():
+        assert _get_instance().get("3", _VALUE) == _VALUE
 
 
 _WRONG_EXPIRY_VALUES = [
-    ("negative number", "-5"),
-    ("not a number", "some string"),
-    ("empty string", ""),
+    pytest.param("-5", id="negative number"),
+    pytest.param("some string", id="not a number"),
+    pytest.param("", id="empty string"),
 ]
 
 
-class CacheInvalidation(TestCase):
-    def test_it_is_set_to_one_day_in_seconds_by_default(self):
+class TestCacheInvalidation:
+    @staticmethod
+    def test_it_is_set_to_one_day_in_seconds_by_default():
         with patch.dict("os.environ", {}):
-            self.assertEqual(_get_instance()._expire_in_seconds, 3600)
+            assert _get_instance()._expire_in_seconds == 3600
 
-    def test_it_can_be_overriden_with_environment_variable(self):
+    @staticmethod
+    def test_it_can_be_overriden_with_environment_variable():
         with patch.dict("os.environ", {TemporaryCache._ENV_VAR_EXPIRY: "10"}):
-            self.assertEqual(_get_instance()._expire_in_seconds, 10)
+            assert _get_instance()._expire_in_seconds == 10
 
-    @parameterized.expand(_WRONG_EXPIRY_VALUES, name_func=get_test_case_name_without_index)
-    def test_it_warns_when_environment_variable_is(self, _name, value):
+    @staticmethod
+    @pytest.mark.parametrize("value", _WRONG_EXPIRY_VALUES)
+    def test_it_warns_when_environment_variable_is(value):
         with patch.dict("os.environ", {TemporaryCache._ENV_VAR_EXPIRY: value}):
-            with self.assertWarnsRegex(RuntimeWarning, ".*improperly configured.*"):
+            with pytest.warns(RuntimeWarning, match=".*improperly configured.*"):
                 _get_instance()
 
-    @parameterized.expand(_WRONG_EXPIRY_VALUES, name_func=get_test_case_name_without_index)
-    def test_it_ignores_invalid_environment_variable_value(self, _name, value):
+    @staticmethod
+    @pytest.mark.parametrize("value", _WRONG_EXPIRY_VALUES)
+    def test_it_ignores_invalid_environment_variable_value(value):
         with patch.dict("os.environ", {TemporaryCache._ENV_VAR_EXPIRY: value}):
-            with self.assertWarnsRegex(RuntimeWarning, ".*Using default value of.*"):
+            with pytest.warns(RuntimeWarning, match=".*Using default value of.*"):
                 _get_instance()
 
-    def test_it_will_not_return_expired_entry(self):
+    @staticmethod
+    def test_it_will_not_return_expired_entry():
         _create_temp_file({_PROJECT: {_KEY_OUT: [_VALUE, 0]}})
-        self.assertIsNone(_get_instance().get(_KEY_IN))
+        assert _get_instance().get(_KEY_IN) is None
 
-    def test_it_will_return_valid_entry(self):
+    @staticmethod
+    def test_it_will_return_valid_entry():
         _create_temp_file({_PROJECT: {_KEY_OUT: [_VALUE, int(time())]}})
-        self.assertEqual(_get_instance().get(_KEY_IN), _VALUE)
+        assert _get_instance().get(_KEY_IN) == _VALUE
 
-    def test_it_doesnt_write_to_cache_when_disabled(self):
+    @staticmethod
+    def test_it_doesnt_write_to_cache_when_disabled():
         _remove_temp_file()
         with patch.dict("os.environ", {TemporaryCache._ENV_VAR_EXPIRY: "0"}):
             _get_instance()[_KEY_IN] = _VALUE
-        self.assertFalse(os.path.isfile(TemporaryCache._TEMP_FILE_NAME))
+        assert not os.path.isfile(TemporaryCache._TEMP_FILE_NAME)
 
-    def test_it_doesnt_read_cache_when_disabled(self):
+    @staticmethod
+    def test_it_doesnt_read_cache_when_disabled():
         _create_temp_file({_PROJECT: {_KEY_OUT: [_VALUE, int(time())]}})
         with patch.dict("os.environ", {TemporaryCache._ENV_VAR_EXPIRY: "0"}):
-            self.assertIsNone(_get_instance().get(_KEY_IN))
+            assert _get_instance().get(_KEY_IN) is None
 
-    @parameterized.expand(_WRONG_EXPIRY_VALUES, name_func=get_test_case_name_without_index)
-    def test_it_ignores_invalid_expiry_cache_value_(self, _name, value):
+    @staticmethod
+    @pytest.mark.parametrize("value", _WRONG_EXPIRY_VALUES)
+    def test_it_ignores_invalid_expiry_cache_value_(value):
         _create_temp_file({_PROJECT: {_KEY_OUT: [_VALUE, value]}})
-        self.assertIsNone(_get_instance().get(_KEY_IN))
+        assert _get_instance().get(_KEY_IN) is None