+++ CHANGELOG.md
@@ -14,6 +14,24 @@ Types of changes are:
 
 ## Unreleased
 
+## Added
+* Compatibility with pytest
+* Repository ID supplied through constructor as a single string
+
+### Changed
+* Tests rewritten from unittest to pytest
+* GitHub issue class and assertions drop duplicate or unnecessary information for their name:
+    * `GitHubIssueTestCase` -> `AssertGitHubIssue`
+    * `assert_github_issue_is_state` -> `is_state`
+    * `assert_github_issue_is_open` -> `is_open`
+    * `assert_github_issue_is_closed` -> `is_closed`
+    * `assert_no_new_release_is_available` -> `current_release`
+* Method parameter `issue_number` renamed to `issue_id` in `is_state()`, `is_open()` and `is_closed()`.
+
+### Removed
+* Sub-classing of `unittest.TestCase`
+* Setting repository ID through class attributes
+
 ## [1.2.0] - 2019-08-14
 
 ### Added